dom/tests/mochitest/mouselock/file_nestedFullScreen.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Test for Bug 633602</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <!--<script type="text/javascript" src="mouselock_util.js"></script>-->
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="start();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
  <p id="display"></p>
  <div id="content" style="display: none">
  </div>

	<div id="parentDiv">
		<div id="childDiv">
		</div>
	</div>

  <pre id="test">
    <script type="application/javascript">
      /** Test for Bug 633602 **/
      SimpleTest.waitForExplicitFinish();

      var pointer = navigator.pointer;
			var parentDiv;
			var childDiv;
			var finished = false;
	
			SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only",
                                  false);	

      document.addEventListener("mozfullscreenchange", function() {
				console.log("mozfullscreenchange");
				if (finished) {
					console.log("finished");
					document.mozCancelFullScreen();
          SimpleTest.finish();
       	} else if (document.mozFullScreenElement === parentDiv) {
					console.log("mozFullScreenElement == parentDiv");
					pointer.lock(parentDiv, function () {
						console.log("parentDiv success callback");
						ok (pointer.islocked, "Mouse should be locked to parentDiv");
						childDiv.mozRequestFullScreen();
					});
        } else if (document.mozFullScreenElement === childDiv) {
					console.log("mozFullScreenElement === childDiv");
					is(pointer.islocked(), false, "Mouse shouldn't be locked to childDiv");
					finished = true;
					document.mozCancelFullScreen();
				}
      }, false);

      function start() {
				console.log("starting...");
				parentDiv = document.getElementById("parentDiv");
				childDiv = document.getElementById("childDiv");

        SimpleTest.waitForFocus(function() {
          parentDiv.mozRequestFullScreen();
        });
      }
    </script>
  </pre>
</body>
</html>
